% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_fpl_val.R
\name{append_fpl_val}
\alias{append_fpl_val}
\title{Generate funnel plot values and output as data table}
\usage{
append_fpl_val(
  numerator,
  denominator,
  establishment,
  funnelcharttype = "PR",
  multiplier = 1,
  betteris,
  period_end
)
}
\arguments{
\item{numerator}{A numeric vector}

\item{denominator}{A numeric vector}

\item{establishment}{A vector of establishment ids}

\item{funnelcharttype}{A string identifying the type of funnel plot.
Default "PR"}

\item{multiplier}{A string identifying the multiplication factor. Default 1}

\item{betteris}{A character string, shows direction of positive change.}

\item{period_end}{A string of the maximum date of the funnel period for use
as the identifier for funnel astronomical points
type Date format: \%d/\%m/\%Y:15/12/2022}
}
\value{
The input data with appended funnel plot values in a data table
}
\description{
Generate funnel plot values and output as data table
}
\examples{
-
}
\concept{pattern detection functions}
