% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Append_spc_val.R
\name{appendSpcVal}
\alias{appendSpcVal}
\title{Append spc chart values to data frame}
\usage{
appendSpcVal(
  numerator,
  denominator,
  period_end,
  spccharttype = "p",
  multiplier = 1
)
}
\arguments{
\item{numerator}{A numeric vector}

\item{denominator}{A numeric vector}

\item{period_end}{A vector of dates of type Date format: \%d/\%m/\%Y:12/11/2021}

\item{spccharttype}{A string identifying the type of spc chart. Default "p"}

\item{multiplier}{A string identifying the multiplication factor. Default 1}
}
\value{
The input data with appended spc data
}
\description{
Append spc chart values to data frame
}
\examples{
-
}
\seealso{
Other Pattern detection functions: 
\code{\link{pattern_detection}()},
\code{\link{pattern_rules}()},
\code{\link{recommendation_rules}()},
\code{\link{validSPC}()}
}
\concept{Pattern detection functions}
