% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pattern_detection.R
\name{compare_pattern_detection}
\alias{compare_pattern_detection}
\title{Compare Pattern Detection}
\usage{
compare_pattern_detection(
  pattern_output,
  filepath_patterns_old = paste0("https://wahealthdept.sharepoint.com/:x:/r/",
    "sites/SafetyandQualityIndicatorSetSQuIS/internal/",
    "SQuIS\%20O365\%20Confidential\%20Documents/Patterns/",
    "MostRecentPatternRecognition/", "LastPatternRun.xlsx",
    "?d=w600d2bd92db349bea66c22dd81b10c74&csf=1&web=1&e=ax7H92"),
  tk_sp,
  sheet_name = "Data"
)
}
\arguments{
\item{pattern_output}{A data table produced by \code{\link[=pattern_detection]{pattern_detection()}}}

\item{filepath_patterns_old}{The SharePoint file path of the last pattern run}

\item{tk_sp}{The SharePoint connection token from \code{\link[qiverse.azure:get_az_tk]{qiverse.azure::get_az_tk()}}}

\item{sheet_name}{The excel sheet name to input}
}
\value{
A combined data table of the old and new patterns
}
\description{
Compare Pattern Detection
}
\examples{
-
}
