% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_key_vault.R
\name{.secrets_operation}
\alias{.secrets_operation}
\title{A generic implementation function for interacting with Azure Key Vault
secrets. Handles the connection and authentication (if necessary) to the
Key Vault service}
\usage{
.secrets_operation(vault_name, operation, token_object, ...)
}
\arguments{
\item{vault_name}{The name of the target Azure Key Vault}

\item{operation}{The name of the operation to be applied, see the
AzureKeyVault::secrets documentation for a full list of possible operations}

\item{token_object}{(optional) A previously-generated AzureAuth token object
created for the Key Vault resource}

\item{...}{Any arguments to be passed to the Key Vault operation, see the
AzureKeyVault::secrets documentation for the possible arguments for each
operation}
}
\value{
An AzureKeyVault::key_vault() object
}
\description{
A generic implementation function for interacting with Azure Key Vault
secrets. Handles the connection and authentication (if necessary) to the
Key Vault service
}
\examples{
-
}
\seealso{
Other Key Vault methods: 
\code{\link{.init_key_vault}()},
\code{\link{kv_delete_secret}()},
\code{\link{kv_get_secret}()},
\code{\link{kv_list_secrets}()},
\code{\link{kv_write_secret}()}
}
\concept{Key Vault methods}
