% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerbi_dataflow.R
\name{download_dataflow_table}
\alias{download_dataflow_table}
\title{Download PowerBI Dataflow Table}
\usage{
download_dataflow_table(
  workspace_name,
  dataflow_name,
  table_name,
  access_token
)
}
\arguments{
\item{workspace_name}{The PowerBI workspace name.}

\item{dataflow_name}{The name of the PowerBI Dataflow within the workspace.}

\item{table_name}{The name of the table within the PowerBI Dataflow to be
accessed.}

\item{access_token}{The token generated with the correct PowerBI Dataflow
permissions. Use get_az_tk('pbi_df') to create this token.}
}
\value{
A tibble of the PowerBI table from the PowerBI dataflow.
}
\description{
This function provides the ability to download tables from within PowerBI
dataflow themselves, rather than requiring that they first be exposed via a
PowerBI dataset.
}
\examples{
 \dontrun{
# Create PowerBI Dataflow azure token
tk <- get_az_tk('pbi_df')

# Load AD dummy file from PowerBI dataflow
ad_dummy <- download_dataflow_table(
  workspace_name = "DOH - Safety and Quality",
  dataflow_name = "active_directory",
  table_name = "dummy",
  access_token = tk$credentials$access_token)
}
}
