% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerbi_dataset_xmla.R
\name{download_dataset_table_com}
\alias{download_dataset_table_com}
\title{Download PowerBI Dataset Table}
\usage{
download_dataset_table_com(
  workspace_name,
  dataset_name,
  table_name,
  access_token,
  verbose = TRUE
)
}
\arguments{
\item{workspace_name}{The PowerBI workspace name.}

\item{dataset_name}{The name of the PowerBI Dataset within the workspace.}

\item{table_name}{The name of the table within the PowerBI Dataset to be
accessed.}

\item{access_token}{The token generated with the correct PowerBI Dataset
permissions. Use get_az_tk('pbi_ds') to create this token.}

\item{verbose}{Print status messages during connection and download}
}
\value{
A tibble of the PowerBI table from the PowerBI dataset.
}
\description{
This function uses Microsoft's Analysis Services client libraries to connect
to a PowerBI dataset by treating it as a SQL Server Analysis Cube. These
libraries should be installed by SQL Server and PowerBI, but if you encounter
issues they can be manually installed from:
https://docs.microsoft.com/en-us/analysis-services/client-libraries
}
\details{
This function uses the Common Object Model (COM) in Windows to connect R
with the client libraries. Consequently, this function is only compatible
with Windows environments
}
\examples{
 \dontrun{
# Create PowerBI Dataset azure token
tk <- get_az_tk('pbi_ds')

# Load AD dummy file from PowerBI dataflow
ad_dummy <- download_dataset_table_com(
  workspace_name = "DOH - Reference Data",
  dataset_name = "r_datasets",
  table_name = "COUNT_lbw",
  access_token = tk$credentials$access_token)
}
}
