% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerbi_dataset_xmla.R
\name{download_dataset_table_pwsh}
\alias{download_dataset_table_pwsh}
\title{This function uses Microsoft's Analysis Services client libraries to connect
to a PowerBI dataset by treating it as a SQL Server Analysis Cube. This
function uses the .NET Framework implementations of Microsoft's Client
Libraries via powershell. Powershell and the .NET libraries are able to be
used on Linux and MacOS via Powershell Core and .NET Core.}
\usage{
download_dataset_table_pwsh(
  workspace_name,
  dataset_name,
  table_name,
  access_token
)
}
\arguments{
\item{workspace_name}{The PowerBI workspace name.}

\item{dataset_name}{The name of the PowerBI Dataset within the workspace.}

\item{table_name}{The name of the table within the PowerBI Dataset to be
accessed.}

\item{access_token}{The token generated with the correct PowerBI Dataset
permissions. Use get_az_tk('pbi_ds') to create this token.}
}
\value{
A tibble of the PowerBI table from the PowerBI dataset.
}
\description{
This function uses Microsoft's Analysis Services client libraries to connect
to a PowerBI dataset by treating it as a SQL Server Analysis Cube. This
function uses the .NET Framework implementations of Microsoft's Client
Libraries via powershell. Powershell and the .NET libraries are able to be
used on Linux and MacOS via Powershell Core and .NET Core.
}
\examples{
 \dontrun{
# Create PowerBI Dataset azure token
tk <- get_az_tk('pbi_ds')

# Load AD dummy file from PowerBI dataset
ad_dummy <- download_dataset_table_pwsh(
  workspace_name = "DOH - Reference Data",
  dataset_name = "r_datasets",
  table_name = "COUNT_lbw",
  access_token = tk$credentials$access_token)
}
}
