% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_list.R
\name{download_sharepoint_list}
\alias{download_sharepoint_list}
\title{Download SharePoint list}
\usage{
download_sharepoint_list(list_url, token)
}
\arguments{
\item{list_url}{The URL of the list when accessing the list via your web
browser.}

\item{token}{The token generated with the correct SharePoint permissions.
Use get_az_tk('sp") to create this token.}
}
\value{
A data.table of the SharePoint list
}
\description{
Download SharePoint list
}
\examples{
 \dontrun{
# Get the token for SharePoint
tk <- get_az_tk('sp')

# Download a particular view for the SharePoint list
list_data <- download_sharepoint_list(
  paste0("https://wahealthdept.sharepoint.com/sites/",
  "SafetyandQualityIndicatorSetSQuIS/Lists/Indicators/",
  "Currently\%20Live.aspx"),
  token = tk
)

# Download the default view for the SharePoint list
list_data <- download_sharepoint_list(
  paste0("https://wahealthdept.sharepoint.com/sites/",
  "SafetyandQualityIndicatorSetSQuIS/Lists/Indicators"),
  token = tk
)
}
}
