% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotly_funnel.R
\name{fpl_plotly_create}
\alias{fpl_plotly_create}
\title{Draw a funnel plot in plotly}
\usage{
fpl_plotly_create(
  numerator,
  denominator,
  group,
  data_type = "PR",
  multiplier = 1,
  betteris = "Higher",
  title = "",
  group_name = NULL,
  short_group_name = NULL,
  parent_group_name = NULL,
  parent_group_name_label = NA,
  funnel_period_start = NA,
  funnel_period_end = NA,
  y_limit_lower = NA,
  y_limit_upper = NA,
  x_axis_label = NA,
  y_axis_label = NA,
  highlight_hosp = NULL,
  highlight_outlier = TRUE,
  highlight_outlier_options = list(direction_to_flag = "Deterioration"),
  brand_colour = "#00667B",
  actual_colour = "black",
  annotation_marker_colour = "grey",
  line_width = 3,
  marker_size = 8,
  y_dp = 1,
  y_format = "Percentage",
  pattern_text_ay = 50,
  nhs_colours_enable = TRUE,
  nhs_colours_options = list(improvement_direction = betteris, direction_to_flag =
    "Both", colours = list(neutral = "#490092", improvement = "#00B0F0", deterioration =
    "#E46C0A", common_cause = "#A6A6A6")),
  show_legend = FALSE,
  source_text = "Healthcare Quality Intelligence Unit"
)
}
\arguments{
\item{numerator}{A vector of the numerator (observed events/counts) values.
Used as the numerator of the Y-axis}

\item{denominator}{A vector of denominator (predicted/population etc.) Used
as denominator of the Y-axis and the scale of the x-axis}

\item{group}{A vector of group names as character or factor. Used to
aggregate and group points on plots}

\item{data_type}{A string identifying the type of data used for in the plot,
the adjustment used and the reference point. One of: "SR" for indirectly
standardised ratios such as HSMR, "PR" for proportions, or "RC" for ratios
of counts. Default is "PR".}

\item{multiplier}{Scale relative risk and funnel by this factor. Default to
1, but 100 sometime used, e.g. in some hospital mortality ratios.}

\item{betteris}{A string identifying the direction that is favourable for
the indicator. "Higher" for points below the lower control limit to be
unfavourable, "Lower" for points above the upper control limit to be
unfavourable, and "Neutral" if the direction is not stated. Default is
"Higher".}

\item{title}{A string for the title of the plot.}

\item{group_name}{A vector of the group names that may differ from the
initial group identifiers. I.e. group is establishment code, and group_name
is the short hospital name.}

\item{short_group_name}{A vector of shortened group names to be displayed
for highlighted hospitals and outlier points.}

\item{parent_group_name}{A vector of parent group names which are to be
displayed in the tooltip. These are the major categories for the groups, i.e.
HSP names.}

\item{parent_group_name_label}{A string for the tooltip label on the parent
group name.}

\item{funnel_period_start}{A date (or character of format #' "yyyy-mm-dd")
for the start date of the funnel period.}

\item{funnel_period_end}{A date (or character of format #' "yyyy-mm-dd")
for the end date of the funnel period.}

\item{y_limit_lower}{A value denoting the y-axis lower limit range}

\item{y_limit_upper}{A value denoting the y-axis upper limit range}

\item{x_axis_label}{A value denoting the x-axis label. Usually NA or a specified label.}

\item{y_axis_label}{A value denoting the y-axis label. Usually NA, a specified label or "Rate",
for items which are proportion based with multipliers.}

\item{highlight_hosp}{Optional hospital name(s) to highlight individually}

\item{highlight_outlier}{Boolean, if FALSE do not highlight outliers}

\item{highlight_outlier_options}{Options to apply if highlight_outlier is TRUE.
Default is list(direction_to_flag = "Deterioration"), where direction_to_flag
can be "Deterioration", "Improvement" or "Both".}

\item{brand_colour}{Hex code for the colour of the funnel limits. Default is
"#00667b"}

\item{actual_colour}{Hex code for the colour actual data points. Default is
"black"}

\item{annotation_marker_colour}{Hex code for the colour actual data points.
Default is "grey"}

\item{line_width}{A value denoting the width of the line in pixels
(default = 3)}

\item{marker_size}{A value denoting the size of the markers in pixels
(default = 8)}

\item{y_dp}{Number of decimal points for the y-axis ticks and hoverinfo.
Default is 1}

\item{y_format}{Set to either "Numeric" or "Percentage" to format the y-axis.
Default is "Percentage"}

\item{pattern_text_ay}{Set the y offset for pattern detection text in pixels
(default = 50)}

\item{nhs_colours_enable}{A boolean to enable NHS colours for the SPC chart.
(default = TRUE)}

\item{nhs_colours_options}{A list of parameters to enable NHS colours for the SPC
chart. (default = list(improvement_direction = betteris,
direction_to_flag = "Both", colours = list(neutral = "#490092",
improvement = "#00B0F0", deterioration = "#E46C0A", common_cause = "#A6A6A6"))}

\item{show_legend}{A boolean to enable legend for the Funnel Plot
(default = FALSE)}

\item{source_text}{Set source text of the chart. If empty ("") or NA, no
source will be displayed (default = "Healthcare Quality Intelligence Unit")}
}
\value{
Funnel plot in plotly for the specified indicator
\subsection{Additional requirements for this function}{

The group vector must be a vector of unique values that are associated with
a numerator and a denominator. Any summation of numerators and denominators
by group must be done prior to input into this function.
}
}
\description{
Draw a funnel plot in plotly
}
\examples{
-
\dontrun{
# Minimal Example

fpl_plotly_create(
  numerator = (sapply(1:5, function(x) rpois(5, x * 20 * 0.6)) |> c() -
                 (runif(25) * 10) |> round(0)) |>
    sapply(function(x) ifelse(x<0, 0, x)),
  denominator = sapply(1:5, function(x) rpois(5, x * 20)) |> c(),
  group = paste0('Q', 1:25),
  data_type = 'PR',
  multiplier = 1,
  betteris = "Higher",
  title = "Test Indicator",
  y_format = "Percentage",
  x_axis_label = "This is X Axis",
  y_axis_label = "This is Y Axis"
)

# Full Example
fpl_plotly_create(
  numerator = (sapply(1:5, function(x) rpois(5, x * 20 * 0.6)) |> c() -
                 (runif(25) * 10) |> round(0)) |>
    sapply(function(x) ifelse(x<0, 0, x)),
  denominator = sapply(1:5, function(x) rpois(5, x * 20)) |> c(),
  group = paste0('Q', 1:25),
  data_type = 'PR',
  multiplier = 1,
  betteris = "Higher",
  title = "Test Indicator",
  group_name = c("Alfa", "Bravo", "Charlie", "Delta", "Echo", "Foxtrot",
                 "Golf","Hotel", "India", "Juliett", "Kilo", "Lima", "Mike",
                 "November","Oscar", "Papa", "Quebec", "Romeo", "Sierra",
                 "Tango", "Uniform", "Victor", "Whiskey", "X-ray", "Yankee"),
  short_group_name = LETTERS[1:25],
  parent_group_name = "Code Words",
  parent_group_name_label = "Category",
  funnel_period_start = '2021-07-01',
  funnel_period_end = '2022-06-30',
  y_limit_lower = NA,
  y_limit_upper = 85,
  x_axis_label = "This is X Axis",
  y_axis_label = "This is Y Axis",
  highlight_hosp = c('Q2', 'Q5'),
  highlight_outlier = TRUE,
  nhs_colours_enable = TRUE,
  show_legend = FALSE,
  y_format = "Percentage",
  source_text = 'Healthcare Quality Intelligence Unit'
)
}
}
