% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{funnel_data}
\alias{funnel_data}
\title{Funnel Data}
\format{
\subsection{\code{funnel_data}}{

A data frame with 66 rows and 4 columns:
\describe{
\item{indicator}{The indicator name that matches the indicator column from
the indicator_data}
\item{group}{The group (hospital) for which this row contains data from}
\item{numerator}{The numerator value for the indicator and hospital
combination}
\item{denominator}{The denominator value for the indicator and hospital
combination}
}
}
}
\usage{
funnel_data
}
\description{
Example funnel plot data used as an input to the data cleaning step for the
Multiple Indicator Sigma Chart.
}
\keyword{datasets}
