% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_data}
\alias{indicator_data}
\title{Indicator Data}
\format{
\subsection{\code{indicator_data}}{

A data frame with 6 rows and 5 columns:
\describe{
\item{indicator_theme}{The theme (group) that the indicator belongs to}
\item{indicator}{The name of the indicator itself}
\item{multiplier}{The multiplier applied to the indicator}
\item{data_type}{The type of funnel plot used in the FunnelPlotR package}
\item{betteris}{The direction of favourability for the indicator. I.e. set
to "Lower" when the indicator is improving as it decreases, and set to
"Higher" when the indicator is improving as it increases.}
}
}
}
\usage{
indicator_data
}
\description{
Example set of indicators used as an input to the data cleaning step for the
Multiple Indicator Sigma Chart.
}
\keyword{datasets}
