% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks_secrets.R
\name{init_databricks_access_token}
\alias{init_databricks_access_token}
\title{Create an Azure authentication token and store it as a Databricks secret
when run on a databricks cluster}
\usage{
init_databricks_access_token(url, username)
}
\arguments{
\item{url}{The workspace URL of the databricks instance}

\item{username}{The current user's username, for the secret scope}
}
\description{
Create an Azure authentication token and store it as a Databricks secret
when run on a databricks cluster
}
\examples{
-
}
\seealso{
Other Azure methods: 
\code{\link{az_authenticated_api_query}()},
\code{\link{db_secret_scopes_api}()},
\code{\link{db_secrets_api}()}
}
\concept{Azure methods}
