% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc}
\alias{misc}
\title{Generate a Multi-Indicator Sigma Chart in Plotly for a Chosen Group}
\usage{
misc(
  funnel_data,
  indicator_data,
  chosen_group,
  hover_colour = "#00667B",
  control_colour = "#00667B",
  worse_colour = "#E46C0A",
  neutral_colour = "#A6A6A6",
  better_colour = "#00B0F0",
  y_dp = 2
)
}
\arguments{
\item{funnel_data}{A data.frame/data.table of the funnel data with the
columns indicator, group, numerator and denominator. See ?example_funnel_data
for an example and more information on the fields required.}

\item{indicator_data}{A data.frame/data.table of the indicator metadata with
the columns indicator_theme, indicator, multiplier, data_type and betteris.
See ?example_indicator_data for an example and more information on the fields
required.}

\item{chosen_group}{A string of the group to be displayed on the chart.}

\item{hover_colour}{The background colour for the hoverinfo (default =
"#00667B")}

\item{control_colour}{The colour of the 95\% and 99\% control limits for the plot
(default = "#00667B")}

\item{worse_colour}{The colour of the bar in the unfavourable direction
(default = "#E46C0A")}

\item{neutral_colour}{The colour of the bar in the favourable direction
(default = "#A6A6A6")}

\item{better_colour}{The colour of the bar in the favourable direction
(default = "#00B0F0")}

\item{y_dp}{The number of decimal places displayed on the chart (default = 2)}
}
\value{
A data.table with the required fields for the MISC plotly chart.
}
\description{
This function generates a Multi-Indicator Sigma Chart in Plotly for a chosen
group. The function requires the funnel data and indicator metadata to be
provided as data.frames/data.tables. The function will then prepare the data
for the MISC chart and plot the chart in Plotly.
}
\examples{
-
\dontrun{
  library(qiverse.qimisc)
  misc_data <- misc(
    funnel_data = example_funnel_data,
    indicator_data = example_indicator_data,
    chosen_group = "A"
  )
}

}
