% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_plotly.R
\name{misc_plotly}
\alias{misc_plotly}
\title{Chart the Multiple Indicator Sigma Chart in Plotly}
\usage{
misc_plotly(
  data,
  brand_colour = "#00667B",
  worse_colour = "black",
  better_colour = "grey",
  y_dp = 2
)
}
\arguments{
\item{data}{A data.frame/data.table of the funnel data with the
columns indicator, group, numerator and denominator. See ?example_funnel_data
for an example and more information on the fields required.}

\item{brand_colour}{The colour of the 95\% and 99\% control limits for the plot
(default = "#00667B")}

\item{worse_colour}{The colour of the bar in the unfavourable direction
(default = "black")}

\item{better_colour}{The colour of the bar in the favourable direction
(default = "grey")}

\item{y_dp}{The number of decimal places displayed on the chart (default = 2)}
}
\value{
A data.table with the required fields for the MISC plotly chart.
}
\description{
Chart the Multiple Indicator Sigma Chart in Plotly
}
\examples{
-
\dontrun{
  library(qiverse.qimisc)
  misc_data <- misc_prep_data(example_funnel_data, example_indicator_data)
  misc_plotly(misc_data[GROUP == "A"])
}
}
