% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_detection.R
\name{pattern_detection}
\alias{pattern_detection}
\title{Pattern Detection}
\usage{
pattern_detection(
  indicator,
  establishment,
  period_end,
  numerator,
  denominator,
  multiplier,
  betteris,
  spccharttype = "p",
  funnelcharttype = "PR",
  indicatorgroup = NA,
  descriptionshort = indicator,
  shorthospitalname = establishment,
  funneldatapoints = "Yes"
)
}
\arguments{
\item{indicator}{A vector of indicator codes}

\item{establishment}{A vector of establishment ids}

\item{period_end}{A vector of dates of type Date format: \%d/\%m/\%Y:12/11/2021}

\item{numerator}{A numeric vector}

\item{denominator}{A numeric vector}

\item{multiplier}{A string identifying the multiplication factor. Default 1}

\item{betteris}{A character string, shows direction of positive change.}

\item{spccharttype}{A string identifying the type of spc chart. Default "p"}

\item{funnelcharttype}{A string identifying the type of funnel plot.
Default "PR"}

\item{indicatorgroup}{A vector used for indicator grouping i.e. QSG Theme}

\item{descriptionshort}{A vector of descriptive names for indicators}

\item{shorthospitalname}{A vector of descriptive names for establishments}

\item{funneldatapoints}{A vector of which data points are included in
the funnel plot calculation}
}
\value{
A data table of unique indicator establishment rows with any
identified patterns and qsg recommendations
}
\description{
Pattern Detection
}
\examples{
-
}
\seealso{
Other Pattern detection functions: 
\code{\link{pattern_rules}()},
\code{\link{recommendation_rules}()},
\code{\link{validSPC}()}
}
\concept{Pattern detection functions}
