% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recommendation_rules.R
\name{recommendation_rules}
\alias{recommendation_rules}
\title{Recommendation rules}
\usage{
recommendation_rules(
  indicator,
  hospital,
  numerator,
  denominator,
  betteris,
  astro,
  shift,
  trend,
  twointhree,
  fpl_astro,
  indicatorgroup = NA
)
}
\arguments{
\item{indicator}{A character vector}

\item{hospital}{A character vector}

\item{numerator}{A numeric vector, sum of the numerator field}

\item{denominator}{A numeric vector, sum of the denominator field}

\item{betteris}{A character string, shows direction of positive change.
"Higher" or "Lower"}

\item{astro}{Date of latest instance of Astronomical Point pattern}

\item{shift}{Date of latest instance of Shift pattern}

\item{trend}{Date of latest instance of Trend pattern}

\item{twointhree}{Date of latest instance of Two in Three pattern}

\item{fpl_astro}{Date of max period end of funnel plot data if outlier}

\item{indicatorgroup}{A character vector, grouping for indicators.
Dropped by default}
}
\value{
A data table with added qsg recommendation and reason
\subsection{Additional requirements for this function}{

The input dataframe must have been passed through \code{\link[rSQuIS:pattern_rules]{rSQuIS::pattern_rules()}}
}
}
\description{
Recommendation rules
}
\examples{
-
}
\seealso{
Other Pattern detection functions: 
\code{\link{appendSpcVal}()},
\code{\link{pattern_detection}()},
\code{\link{pattern_rules}()},
\code{\link{validSPC}()}
}
\concept{Pattern detection functions}
