% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_detection_stripped.R
\name{runFPat}
\alias{runFPat}
\title{Generate funnel plot pattern detection output over input data}
\usage{
runFPat(
  numerator,
  denominator,
  establishment,
  funnelcharttype,
  multiplier,
  betteris
)
}
\arguments{
\item{numerator}{A numeric vector}

\item{denominator}{A numeric vector}

\item{establishment}{A character vector}

\item{funnelcharttype}{A character string}

\item{multiplier}{A numeric value}

\item{betteris}{A character string, either "Lower" or "Higher"}
}
\value{
A dataframe containing the outliers for the funnel plot
}
\description{
Generate funnel plot pattern detection output over input data
}
\examples{
-
}
