% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_detection_stripped.R
\name{runFPatterns}
\alias{runFPatterns}
\title{Generate pattern detection output for Funnel plots}
\usage{
runFPatterns(
  input_df = NULL,
  numerator = NULL,
  denominator = NULL,
  funnelcharttype = NULL,
  multiplier = NULL,
  betteris = NULL,
  indicator = NULL,
  establishment = NULL
)
}
\arguments{
\item{input_df}{A dataframe filtered to the squis format.}

\item{numerator}{A numeric vector}

\item{denominator}{A numeric vector}

\item{funnelcharttype}{A character string}

\item{multiplier}{A numeric value}

\item{betteris}{A character string, either "Lower" or "Higher"}

\item{indicator}{A character string}

\item{establishment}{A character vector}
}
\value{
A dataframe consisting of the pattern detection output with
added indicator and establishment
}
\description{
Generate the pattern detection output for the input data.
Can handle either input_df formatted to the squis format or individual
vectors for numerator, denominator, period_end and indicator, and strings for
funnelcharttype, multiplier, betteris and establishment.

Wrapper for the \code{\link[=runFPat]{runFPat()}} function
}
\examples{
-
}
