% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_detection_stripped.R
\name{runPat}
\alias{runPat}
\title{Generate SPC chart pattern detection output over input data}
\usage{
runPat(
  numerator,
  denominator,
  period_end,
  spccharttype,
  multiplier,
  betteris,
  trend_size = 5,
  shift_size = 7
)
}
\arguments{
\item{numerator}{A numeric vector}

\item{denominator}{A numeric vector}

\item{period_end}{A character vector}

\item{spccharttype}{A character string}

\item{multiplier}{A numeric value}

\item{betteris}{A character string, either "Lower" or "Higher"}

\item{trend_size}{A numeric value for the number of points for a trend
pattern (default = 5)}

\item{shift_size}{A numeric value for the number of points for a shift
pattern (default = 7)}
}
\value{
A dataframe containing the most recent pattern for each pattern type
}
\description{
Generate SPC chart pattern detection output over input data
}
\examples{
runPat(
  numerator = c(1,4,3,6,3),
  denominator = c(2,4,6,8,3),
  period_end = c("2022/01/01", "2022/02/01", "2022/03/01",
  "2022/04/01", "2022/05/01"),
  spccharttype = "p",
  multiplier = 100,
  betteris = "Lower",
  trend_size = 5,
  shift_size = 7
)
}
