% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_detection_stripped.R
\name{runPatterns}
\alias{runPatterns}
\title{Generate pattern detection output for
Statistical Process Control charts}
\usage{
runPatterns(
  input_df = NULL,
  numerator = NULL,
  denominator = NULL,
  period_end = NULL,
  spccharttype = NULL,
  multiplier = NULL,
  betteris = NULL,
  indicator = NULL,
  establishment = NULL,
  trend_size = 5,
  shift_size = 7
)
}
\arguments{
\item{input_df}{A dataframe filtered to the squis format.}

\item{numerator}{A numeric vector}

\item{denominator}{A numeric vector}

\item{period_end}{A character vector}

\item{spccharttype}{A character string}

\item{multiplier}{A numeric value}

\item{betteris}{A character string, either "Lower" or "Higher"}

\item{indicator}{A character string}

\item{establishment}{A character string}

\item{trend_size}{A numeric value for the number of points for a trend
pattern (default = 5)}

\item{shift_size}{A numeric value for the number of points for a shift
pattern (defualt = 7)}
}
\value{
A dataframe consisting of the pattern detection output with
added indicator and establishment
}
\description{
Generate the pattern detection output for the input data.
Can handle either input_df formatted to the squis format or individual
vectors for numerator, denominator and period_end, and strings for
spccharttype, multiplier, betteris, indicator and establishment.

Wrapper for the \code{\link[=runPat]{runPat()}} function
}
\examples{
test <- runPatterns(numerator = c(1,4,3,6,3),
                    denominator = c(2,4,6,8,3),
                    period_end = c("2022/01/01", "2022/02/01",
                    "2022/03/01","2022/04/01", "2022/05/01"),
                    spccharttype = "p",
                    multiplier = 100,
                    betteris = "Lower",
                    indicator = "Test Indicator",
                    establishment = "Test Est")
}
