% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_funnel_matrix_prep_data.R
\name{spc_funnel_matrix_prep_data}
\alias{spc_funnel_matrix_prep_data}
\title{Format data for the SPC Funnel Matrix Visuals}
\usage{
spc_funnel_matrix_prep_data(
  indicator,
  establishment,
  period_end,
  period_start,
  numerator,
  denominator,
  multiplier = 1,
  betteris = "Higher",
  parent_group_name = "All",
  indicatorgroup = "All",
  spccharttype = "p",
  funnelcharttype = "PR",
  descriptionshort = indicator,
  shorthospitalname = establishment,
  funneldatapoints = "Yes"
)
}
\arguments{
\item{indicator}{A vector to denote the indicator group.
Used to calculate aggregate SPC and Funnel Plots}

\item{establishment}{A vector of group names.
Used to aggregate and group points on plots}

\item{period_end}{A date (or character of format "yyyy-mm-dd")
for the end date of the period.}

\item{period_start}{A date (or character of format "yyyy-mm-dd")
for the start date of the period.}

\item{numerator}{A vector of the numerator. Used to calculate
the value for SPC and Funnel plots}

\item{denominator}{A vector of the denominator Used to calculate
the value for SPC and Funnel plots}

\item{multiplier}{Scale relative risk and funnel by this factor. Default to
1, but 100 sometime used, e.g. in some hospital mortality ratios.}

\item{betteris}{A string identifying the direction that is favourable for
the indicator. "Higher" for points below the lower control limit to be
unfavourable, "Lower" for points above the upper control limit to be
unfavourable, and "Neutral" if the direction is not stated.
Default is "Higher"}

\item{parent_group_name}{A vector of parent group names which are to be
displayed in the tooltip. These are the major categories for the groups, i.e.
HSP names. Default is "All"}

\item{indicatorgroup}{A vector used for indicator grouping i.e. QSG Theme.
Default is "All"}

\item{spccharttype}{A string identifying the type of spc chart. Default "p"}

\item{funnelcharttype}{A string identifying the type of funnel plot.
Default "PR"}

\item{descriptionshort}{A vector of descriptive names for indicator.
Default indicator variable}

\item{shorthospitalname}{A vector of descriptive names for establishments
Default establishment variable}

\item{funneldatapoints}{A vector of "Yes" or NA, marks which data points
are included in the funnel plot calculation. Default "Yes"}
}
\value{
A data.table with the required fields for the SPC Funnel Matrix.
}
\description{
Extract all favourable and unfavourable SPC and Funnel Plot
patterns, and export into the SPC Funnel Matrix format.
}
\examples{
-
\dontrun{
  library(qiverse.qimatrix)
  spc_funnel_matrix_data <- spc_funnel_matrix_prep_data(
    indicator,
    establishment,
    period_end,
    period_start,
    numerator,
    denominator,
    multiplier = 1,
    betteris = "Higher",
    parent_group_name = "All",
    indicatorgroup = "All",
    spccharttype = "p",
    funnelcharttype = "PR",
    descriptionshort = indicator,
    shorthospitalname = establishment,
    funneldatapoints = "Yes"
  )
}
}
