% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotly_spc.R
\name{spc_plotly_create}
\alias{spc_plotly_create}
\title{Draw a statistical process control chart in plotly}
\usage{
spc_plotly_create(
  x,
  numerator,
  denominator = NULL,
  data_type = "p",
  multiplier = 1,
  betteris = "Lower",
  title = "",
  spc_period_start = NA,
  spc_period_end = NA,
  x_axis_label = NA,
  y_axis_label = NA,
  brand_colour = "#00667B",
  actual_colour = "black",
  annotation_marker_colour = "grey",
  line_width = 2,
  marker_size = 10,
  y_dp = 1,
  y_format = "Percentage",
  x_format = "\%b \%Y",
  patterns = "Yes",
  pattern_text_ay = 50,
  source_text = "Healthcare Quality Intelligence Unit"
)
}
\arguments{
\item{x}{A vector of the x-axis for the SPC chart. Usually dates.}

\item{numerator}{A vector of the numerator (observed events/counts) values.
Used as the numerator of the Y-axis}

\item{denominator}{A vector of denominator (predicted/population etc.) Used
as denominator of the Y-axis and the scale of the x-axis. Only required when
when necessary.}

\item{data_type}{A string identifying the type of data used for in the plot,
the adjustment used and the reference point. One of: "i" for indirectly
standardised ratios such as HSMR, "p" for proportions, "r" for ratios
of counts, "run" for run charts, "g" for geometric charts and "t" for time
charts. Default is "p".}

\item{multiplier}{Scale relative risk and funnel by this factor. Default to
1, but 100 sometime used, e.g. in some hospital mortality ratios.}

\item{betteris}{A string identifying the direction that is favourable for
the indicator. "Higher" for points below the lower control limit to be
unfavourable, and "Lower" for points above the upper control limit to be
unfavourable. Default is "Higher".}

\item{title}{A string for the title of the plot.}

\item{spc_period_start}{A date (or character of format #' "yyyy-mm-dd")
for the start date of the SPC period.}

\item{spc_period_end}{A date (or character of format #' "yyyy-mm-dd")
for the end date of the SPC period.}

\item{x_axis_label}{A value denoting the x-axis label. Usually NA or a specified label.}

\item{y_axis_label}{A value denoting the y-axis label. Usually NA, a specified label or "Rate",
for items which are proportion based with multipliers.}

\item{brand_colour}{Hex code for the colour of the funnel limits}

\item{actual_colour}{Hex code for the colour actual data points}

\item{annotation_marker_colour}{Hex code for the colour actual data points}

\item{line_width}{A value denoting the width of the line in pixels
(default = 2)}

\item{marker_size}{A value denoting the size of the markers in pixels
(default = 10)}

\item{y_dp}{Number of decimal points for the y-axis ticks and hoverinfo}

\item{y_format}{Set to either "Numeric" or "Percentage" to format the y-axis}

\item{x_format}{The format for the x-axis. Default is "\%b \%Y".}

\item{patterns}{Either set to "Yes", "No" or as a dataframe with only a
single row with the columns "Astro", "Trend", "TwoInThree", "Shift". This is
the output of the runPat function in qiverse.qipatterns, or a filtered
pattern dataframe. The "Yes" option automatically applies this function to
the data. The "No" option does not display any patterns. (default = "Yes")}

\item{pattern_text_ay}{Set the y offset for pattern detection text in
pixels (default = 50)}

\item{source_text}{Set source text of the chart. If empty ("") or NA, no
source will be displayed (default = "Healthcare Quality Intelligence Unit")}
}
\value{
Statistical process control chart in plotly for the specific
indicator and establishment.

###Additional requirements for this function
NA
}
\description{
Draw a statistical process control chart in plotly
}
\examples{
-
\dontrun{
}
}
