% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validSPC.R
\name{validSPC}
\alias{validSPC}
\title{Check validity of data for SPC}
\usage{
validSPC(numerator, denominator, period_end)
}
\arguments{
\item{numerator}{A numeric vector}

\item{denominator}{A numeric vector}

\item{period_end}{A vector of dates of type Date format: \%d/\%m/\%Y:15/12/2022}
}
\value{
The input data with added validSPC flag for filtering
}
\description{
Check validity of data for SPC
}
\examples{
-

}
\seealso{
Other Pattern detection functions: 
\code{\link{pattern_detection}()},
\code{\link{pattern_rules}()},
\code{\link{recommendation_rules}()}
}
\concept{Pattern detection functions}
