% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_file.R
\name{download_sharepoint_file}
\alias{download_sharepoint_file}
\title{Download SharePoint File}
\usage{
download_sharepoint_file(
  site_url,
  file_url,
  token,
  download = FALSE,
  download_dest = NULL
)
}
\arguments{
\item{site_url}{This is the url when first accessing the sharepoint team via
your web browser (Home page)}

\item{file_url}{This is a direct link to the file to be downloaded. It can be
obtained by navigating to the file you want to download, right-clicking and
selecting "Copy Link"}

\item{token}{The token generated with the correct SharePoint permissions. Use
get_az_tk('sp") to create this token.}

\item{download}{Set this to FALSE to download the file into a temporary file.
This temporary file can then be read in using any standard data reading
function. Set this to TRUE to download the file to your computer.}

\item{download_dest}{This option is only active when download is set to TRUE.
When download_location is set to NULL, the file is just downloaded into the
current working directory. Otherwise, specify the destination path to save
the file.}
}
\value{
A temporary file stored in an object if download is set to FALSE. If
download is set to TRUE, the selected file is downloaded to the chosen
directory.
}
\description{
Download SharePoint File
}
\examples{
 \dontrun{
# Create SharePoint azure token
tk <- get_az_tk('sp')

# Download file to working directory from SharePoint site
download_sharepoint_file(
  site_url = paste0("https://wahealthdept.sharepoint.com/sites/",
  "SafetyandQualityIndicatorSetSQuIS/"),
  file_url = paste0("https://wahealthdept.sharepoint.com/:x:/r/sites/",
  "SafetyandQualityIndicatorSetSQuIS/Shared\%20Documents/",
  "Reference\%20Material/Establishment_CSV_Exports/",
  "Hospitals_SharePoint_list.csv",
  "?d=wa2a4067cf69241ce8f4a5d209878328b&csf=1&web=1&e=ybGCQF"),
  token = tk,
  download = TRUE
)

# Download file to C:/ drive from SharePoint subsite
download_sharepoint_file(
  site_url = paste0("https://wahealthdept.sharepoint.com/sites/",
  "SafetyandQualityIndicatorSetSQuIS/internal/"), # nolint
  file_url = paste0("https://wahealthdept.sharepoint.com/:u:/r/sites/",
  "SafetyandQualityIndicatorSetSQuIS/internal/",
  "SQuIS\%20O365\%20Confidential\%20Documents/Analytical\%20Code/",
  "EmergencyTest.R?csf=1&web=1&e=pwUOBk"),
  token = tk,
  download = TRUE,
  download_dest = 'C:/Github/EmergencyTest.R'
)

# Example to download to tempfile, to load straight into R memory
data <- download_sharepoint_file(
  site_url = paste0("https://wahealthdept.sharepoint.com/sites/",
  "SafetyandQualityIndicatorSetSQuIS/"), # nolint
  file_url = paste0("https://wahealthdept.sharepoint.com/:x:/r/sites/",
  "SafetyandQualityIndicatorSetSQuIS/Shared\%20Documents/",
  "Reference\%20Material/Establishment_CSV_Exports/",
  "Hospitals_SharePoint_list.csv",
  "?d=wa2a4067cf69241ce8f4a5d209878328b&csf=1&web=1&e=ybGCQF"),
  token = tk,
  download = FALSE
) |>
  # Pipe in the approprate function to read in the data
  read.csv()
}
}
