% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_detection.R
\name{pattern_detection}
\alias{pattern_detection}
\title{Pattern Detection}
\usage{
pattern_detection(
  indicator,
  group,
  period_end,
  numerator,
  denominator,
  multiplier,
  better_is,
  spc_chart_type = "p",
  funnel_chart_type = "PR",
  indicator_group = NA,
  description_short = indicator,
  group_name = group,
  funnel_data_points = "Yes",
  trend_size = 5,
  shift_size = 7
)
}
\arguments{
\item{indicator}{A vector of indicator codes}

\item{group}{A vector of group ids}

\item{period_end}{A vector of dates of type Date format: \%d/\%m/\%Y:12/11/2021}

\item{numerator}{A numeric vector}

\item{denominator}{A numeric vector}

\item{multiplier}{A string identifying the multiplication factor. Default 1}

\item{better_is}{A character string, shows direction of positive change.}

\item{spc_chart_type}{A string identifying the type of spc chart. Default "p"}

\item{funnel_chart_type}{A string identifying the type of funnel plot.
Default "PR"}

\item{indicator_group}{A vector used for indicator grouping i.e. QSG Theme}

\item{description_short}{A vector of descriptive names for indicators}

\item{group_name}{A vector of descriptive names for groups}

\item{funnel_data_points}{A vector of which data points are included in
the funnel plot calculation}

\item{trend_size}{The number of points in a trend pattern}

\item{shift_size}{The number of points in a shift pattern}
}
\value{
A data table of unique indicator group rows with any
identified patterns and qsg recommendations
}
\description{
Pattern Detection
}
\examples{
-
}
\seealso{
Other Pattern detection functions: 
\code{\link{pattern_rules}()},
\code{\link{valid_spc}()}
}
\concept{Pattern detection functions}
