% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_funnel_matrix_flextable.R
\name{spc_funnel_matrix_flextable}
\alias{spc_funnel_matrix_flextable}
\title{Construct Flextable from SPC Funnel Matrix Data}
\usage{
spc_funnel_matrix_flextable(data)
}
\arguments{
\item{data}{A data.table output from the spc_funnel_matrix_prep_data
function}
}
\value{
A flextable visualisation for the SPC Funnel Matrix.
}
\description{
Construct a Flextable from the SPC Funnel Matrix Data
}
\examples{
-
\dontrun{
  library(qiverse.qimatrix)
  spc_funnel_matrix_data <- spc_funnel_matrix_prep_data(
    indicator,
    establishment,
    period_end,
    period_start,
    numerator,
    denominator,
    multiplier,
    betteris,
    parent_group_name,
    indicatorgroup,
    spccharttype = "p",
    funnelcharttype = "PR",
    descriptionshort = indicator,
    shorthospitalname = establishment,
    funneldatapoints = "Yes",
    upload_snowflake = "No"
  )
  spc_funnel_matrix_flextable(data = spc_funnel_matrix_data)
}
}
