% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_funnel_matrix_flextable.R
\name{spc_funnel_matrix_flextable}
\alias{spc_funnel_matrix_flextable}
\title{Construct Flextable from SPC Funnel Matrix Data}
\usage{
spc_funnel_matrix_flextable(
  data,
  filter_indicator_group = NULL,
  filter_indicator = NULL,
  filter_parent_group = NULL,
  filter_hospital = NULL
)
}
\arguments{
\item{data}{A data.table output from the spc_funnel_matrix_prep_data
function}

\item{filter_indicator_group}{A vector of indicator_group to filter the data by. Data
will not be filtered by default (NA)}

\item{filter_indicator}{A vector of indicator to filter the data by. Data
will not be filtered by default (NA)}

\item{filter_parent_group}{A vector of parent_group to filter the data by. Data
will not be filtered by default (NA)}

\item{filter_hospital}{A vector of hospital to filter the data by. Data
will not be filtered by default (NA)}
}
\value{
A flextable visualisation for the SPC Funnel Matrix.
}
\description{
Construct a Flextable from the SPC Funnel Matrix Data
}
\examples{
-
\dontrun{
  library(qiverse.qimatrix)
  library(qiverse.data)
  spc_funnel_matrix_data <- spc_funnel_matrix_prep_data(
    indicator = example_spc_data$indicator,
    group = example_spc_data$group,
    period_end = example_spc_data$period_end,
    period_start = example_spc_data$period_start,
    numerator = example_spc_data$numerator,
    denominator = example_spc_data$denominator,
    multiplier = example_spc_data$multiplier,
    better_is = example_spc_data$better_is,
    parent_group_name = "All",
    indicator_group = "All",
    spc_chart_type = example_spc_data$spc_chart_type,
    funnel_chart_type = example_spc_data$funnel_chart_type,
    funnel_data_points = "Yes"
  )
  spc_funnel_matrix_flextable(data = spc_funnel_matrix_data)
}
}
