% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotly_spc.R
\name{spc_plotly_create}
\alias{spc_plotly_create}
\title{Draw a statistical process control chart in plotly}
\usage{
spc_plotly_create(
  x,
  numerator,
  denominator = NULL,
  data_type = "p",
  multiplier = 1,
  betteris = "Lower",
  title = "",
  spc_period_start = NA,
  spc_period_end = NA,
  x_axis_label = NA,
  y_axis_label = NA,
  brand_colour = "#00667B",
  actual_colour = "black",
  annotation_marker_colour = "grey",
  line_width = 2,
  marker_size = 10,
  y_dp = 1,
  y_format = "Percentage",
  x_format = "\%b \%Y",
  patterns = "Yes",
  pattern_text_ay = 50,
  pattern_arrow_width = 2,
  pattern_font_size = 25,
  trend_size = 5,
  shift_size = 7,
  target = NA,
  target_annotation = NA,
  target_options = list(legend_name = "Target", line_colour = "#FF0000", line_width = 2,
    line_dash = "solid"),
  nhs_colours_enable = TRUE,
  nhs_colours_options = list(improvement_direction = betteris, direction_to_flag =
    "Both", colours = list(neutral = "#490092", improvement = "#00B0F0", deterioration =
    "#E46C0A", common_cause = "#A6A6A6")),
  nhs_icons_enable = FALSE,
  nhs_icons_options = list(flag_last_point_only = FALSE, sizex = 0.075, sizey = 0.15),
  show_legend = FALSE,
  legend_font_size = 12,
  source_text = "Healthcare Quality Intelligence Unit"
)
}
\arguments{
\item{x}{A vector of the x-axis for the SPC chart. Usually dates.}

\item{numerator}{A vector of the numerator (observed events/counts) values.
Used as the numerator of the Y-axis}

\item{denominator}{A vector of denominator (predicted/population etc.) Used
as denominator of the Y-axis and the scale of the x-axis. Only required when
when necessary.}

\item{data_type}{A string identifying the type of data used for in the plot,
the adjustment used and the reference point. One of: "i" for indirectly
standardised ratios such as HSMR, "p" for proportions, "r" for ratios
of counts, "run" for run charts, "g" for geometric charts and "t" for time
charts. Default is "p".}

\item{multiplier}{Scale relative risk and funnel by this factor. Default to
1, but 100 sometime used, e.g. in some hospital mortality ratios.}

\item{betteris}{A string identifying the direction that is favourable for
the indicator. "Higher" for points below the lower control limit to be
unfavourable, "Lower" for points above the upper control limit to be
unfavourable, and "Neutral" if the direction is not stated. Default is
"Higher".}

\item{title}{A string for the title of the plot.}

\item{spc_period_start}{A date (or character of format #' "yyyy-mm-dd")
for the start date of the SPC period.}

\item{spc_period_end}{A date (or character of format #' "yyyy-mm-dd")
for the end date of the SPC period.}

\item{x_axis_label}{A value denoting the x-axis label. Usually NA or a specified label.}

\item{y_axis_label}{A value denoting the y-axis label. Usually NA, a specified label or "Rate",
for items which are proportion based with multipliers.}

\item{brand_colour}{Hex code for the colour of the funnel limits}

\item{actual_colour}{Hex code for the colour actual data points}

\item{annotation_marker_colour}{Hex code for the colour actual data points}

\item{line_width}{A value denoting the width of the line in pixels
(default = 2)}

\item{marker_size}{A value denoting the size of the markers in pixels
(default = 10)}

\item{y_dp}{Number of decimal points for the y-axis ticks and hoverinfo}

\item{y_format}{Set to either "Numeric" or "Percentage" to format the y-axis}

\item{x_format}{The format for the x-axis. Default is "\%b \%Y".}

\item{patterns}{Either set to "Yes", "No" or as a dataframe with only a
single row with the columns "Astro", "Trend", "TwoInThree", "Shift". This is
the output of the runPat function in qiverse.qipatterns, or a filtered
pattern dataframe. The "Yes" option automatically applies this function to
the data. The "No" option does not display any patterns. (default = "Yes")}

\item{pattern_text_ay}{Set the y offset for pattern detection text in
pixels (default = 50)}

\item{pattern_arrow_width}{Set the width of the pattern arrow in pixels
(default = 2)}

\item{pattern_font_size}{Numeric value for the font size of the annotation.
Default is 25}

\item{trend_size}{Set the number of points required for a trend to be
detected (default = 5)}

\item{shift_size}{Set the number of points required for a shift to be
detected (default = 7)}

\item{target}{Set a target value for the SPC chart. If NA, no target will be
displayed. (default = NA)}

\item{target_annotation}{A string to annotate the target line with a text on the plot.
A value of NA indicates no annotation. (default = NA)}

\item{target_options}{A list of parameters to set the target line:
\itemize{
\item legend_name: A string to set the legend name for the target line
\item line_colour: A hex code to set the colour of the target line
\item line_width: A numeric value to set the width of the target line
\item line_dash: A string to set the dash of the target line. Options are
"solid", "dot", "dash", "longdash", "dashdot", "longdashdot"
}

(default = list(legend_name = "Target", line_colour = "#FF0000",
line_width = 2, line_dash = "solid"))}

\item{nhs_colours_enable}{A boolean to enable NHS colours for the SPC chart.
(default = TRUE)}

\item{nhs_colours_options}{A list of parameters to enable NHS colours for the SPC
chart. (default = list(improvement_direction = betteris,
direction_to_flag = "Both", colours = list(neutral = "#490092",
improvement = "#00B0F0", deterioration = "#E46C0A", common_cause = "#A6A6A6"))}

\item{nhs_icons_enable}{A boolean to enable NHS icons for the SPC chart. This
option is only available if nhs_colours_enable is TRUE. (default = FALSE)}

\item{nhs_icons_options}{A list of options for the NHS icons:
\itemize{
\item flag_last_point_only: A boolean to enable flagging of the last point only
when determining the NHS icons. (default = FALSE)
\item sizex: A numeric value to set the size of the NHS icons as a proportion of
the plot's width. (default = 0.075)
\item sizey: A numeric value to set the size of the NHS icons as a proportion of
the plot's height. (default = 0.15)
(default = list(flag_last_point_only = FALSE, sizex = 0.075, sizey = 0.15))
}}

\item{show_legend}{A boolean to enable legend for the SPC Chart
(default = FALSE)}

\item{legend_font_size}{A numeric value to set the font size of the legend
(default = 12)}

\item{source_text}{Set source text of the chart. If empty ("") or NA, no
source will be displayed (default = "Healthcare Quality Intelligence Unit")}
}
\value{
Statistical process control chart in plotly for the specific
indicator and establishment.
\subsection{Additional requirements for this function}{

NA
}
}
\description{
Draw a statistical process control chart in plotly
}
\examples{
-
\dontrun{
spc_plotly_create(
  x = seq(from = as.Date('2019-02-01'),
          to = as.Date('2022-01-01'),
          by = 'month')-1,
  numerator = c(20,25,17,22,18,18,30,29,18,23,17,18,
                15,17,19,21,30,19,15,17,22,24,20,13,
                14,18,14,21,17,27,24,30,25,24,26,24),
  denominator = c(86,91,99,110,96,110,96,97,105,87,94,89,
                  102,106,107,95,106,132,95,117,97,99,108,106,
                  101,96,96,98,105,117,94,77,97,90,106,107),
  data_type = 'p',
  multiplier = 1,
  betteris = "Lower",
  title = paste0("Example Indicator", " - ", "Hospital"),
  spc_period_start = "2019-01-01",
  spc_period_end = "2021-12-31",
  x_axis_label = "Date",
  y_axis_label = "Proportion",
  brand_colour = "#00667B",
  actual_colour = "black",
  annotation_marker_colour = "grey",
  y_dp = 1,
  y_format = "Percentage",
  x_format = "\%b \%Y",
  patterns = "Yes",
  pattern_text_ay = 50,
  trend_size = 5,
  shift_size = 7,
  nhs_colours_enable = TRUE,
  show_legend = FALSE,
  legend_font_size = 12,
  source_text = 'Healthcare Quality Intelligence Unit'
)
}
}
